// backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ProviderStatusEnum = {
  APPROVED: 'APPROVED',
  REVOKED: 'REVOKED',
  PENDING: 'PENDING'
} as const

export type ProviderStatusEnum = (typeof ProviderStatusEnum)[keyof typeof ProviderStatusEnum]


export const AIModelType = {
  LANGUAGE_MODEL: 'LANGUAGE_MODEL',
  IMAGE_GENERATION_MODEL: 'IMAGE_GENERATION_MODEL'
} as const

export type AIModelType = (typeof AIModelType)[keyof typeof AIModelType]


export const HumanOrAIEnum = {
  HUMAN: 'HUMAN',
  AI: 'AI'
} as const

export type HumanOrAIEnum = (typeof HumanOrAIEnum)[keyof typeof HumanOrAIEnum]